* Fix check nil storage key from setStorageValue  https://github.com/opentibiabr/canary/pull/2636
When a key from Lua is nil, it is automatically transformed into 0 without being checked. Consequently, even with a nil key, no error is thrown, and the code continues as if everything had occurred normally, resulting in the storage not being set.

Removed two duplicated checks in creature.cpp; they are already verified in the if statement above.