* Optimized moveCreature and spectators
• In moveCreature, a check was added so that the spectator's search is
not called twice, now only one position is passed with the interval
adjusted to assume both positions as if it were.

• In spectators we no longer use vector_set but just vector and if there
is a need to remove record duplication, it will use unordered_set
passing the vector in the constructor which is much faster.

• Why was useCacheMap disabled?
As the map cache is done synchronously for each movement that a monster
makes, it is better to disable it,
as the pathfinder, which is one of the resources that uses this cache
the most,
is multithreding and thus the processing cost is divided between the
threads.

Thanks @mehah